+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.2.5 - Quality of life improvements
+- Added **macOS** support. 
+- Added **Hotlink** support.
+- Added **Re-scan Playlists** & **Re-scan playlist** features to Tab & Footer views.
+- Added **Icons** in drop down selector for easy playlist type identification.
+- Added **YouTube Privacy** toggle for privacy-enhanced domain on embeds to help with Error 153 issues.
+
+
 ## 1.2.5 - First aimed community release
 - First aimed community release of **Audio Pluglist**.
 - Added **External link** support.